[REFACTOR] 타입 및 상수 정의 분리
